backend/pyproject.toml
[project]
name = "hackerton-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "HyeonsuRyu",email = "hsoo112149@gmail.com"}
]
readme = "README.md"
requires-python = "3.12.*"
dependencies = [
    "django (>=5.1,<6.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "openai (>=2.14.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"
